#164157738 Adds Admin Only Role for routes

+++ api/v2/views/parties_view.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, request, jsonify, make_response
 from api.v2.models.parties_model import PartiesModelDb
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from api.v2.models.user_model import UserModelDb
 
 parties_api_v2 = Blueprint('parties_v2', __name__, url_prefix="/api/v2")
 
@@ -8,22 +9,26 @@ parties_api_v2 = Blueprint('parties_v2', __name__, url_prefix="/api/v2")
 @parties_api_v2.route("/parties", methods=['POST'])
 @jwt_required
 def api_create_parties():
-    party = request.get_json(force=True)
-    if {'name', 'hqAddress', 'logoUrl'} <= set(party):
-        party_name = PartiesModelDb(party).create_party()
-        if 'Party Exists' in party_name:
-            return make_response(jsonify({"status": 409, "error": "Party Already Exists"}), 409)
-        elif 'Invalid Data' in party_name:
-            return make_response(jsonify({"status": 400, "error": "Check Input Values"}), 400)
+    current_user = get_jwt_identity()
+    response = UserModelDb().get_user_by_id(current_user)
+    if 'Requires Admin Privilege' not in response:
+        party = request.get_json(force=True)
+        if {'name', 'hqAddress', 'logoUrl'} <= set(party):
+            party_name = PartiesModelDb(party).create_party()
+            if 'Party Exists' in party_name:
+                return make_response(jsonify({"status": 409, "error": "Party Already Exists"}), 409)
+            elif 'Invalid Data' in party_name:
+                return make_response(jsonify({"status": 400, "error": "Check Input Values"}), 400)
 
-        response_body = {
-            "status": 201,
-            "data": [{
-                "name": party_name
-            }]
-        }
-        return make_response(jsonify(response_body), 201)
-    return make_response(jsonify({"status": 400, "error": "Missing Key value"}), 400)
+            response_body = {
+                "status": 201,
+                "data": [{
+                    "name": party_name
+                }]
+            }
+            return make_response(jsonify(response_body), 201)
+        return make_response(jsonify({"status": 400, "error": "Missing Key value"}), 400)
+    return make_response(jsonify({"status": 401, "error": "Unauthorized Access,Requires Admin Rights"}), 401)
 
 
 @parties_api_v2.route("/parties", methods=['GET'])
@@ -36,22 +41,28 @@ def api_get_parties():
 @parties_api_v2.route("/parties/<party_id>/name", methods=['PATCH'])
 @jwt_required
 def api_edit_party(party_id):
-    oid = id_conversion(party_id)
-    updated_party_data = request.get_json(force=True)
-    if {'name'} <= set(updated_party_data):
-        model_result = PartiesModelDb(party_id=oid).edit_party(updated_party_data['name'])
-        if 'Invalid Id' in model_result or 'Invalid Data' in model_result:
-            return make_response(jsonify({"status": 400, "error": "Invalid Data ,Check id or data being updated"}), 400)
-        elif 'Party Exists' in model_result:
-            return make_response(jsonify({"status": 409, "error": "Party with similar name exists"}), 409)
-        return make_response(jsonify({"status": 200, "message": "{} Updated Successfully".format(model_result[0][1])}),
-                             200)
-    return make_response(jsonify({"status": 400, "error": "Missing Key value"}), 400)
+    current_user = get_jwt_identity()
+    response = UserModelDb().get_user_by_id(current_user)
+    if 'Requires Admin Privilege' not in response:
+        oid = id_conversion(party_id)
+        updated_party_data = request.get_json(force=True)
+        if {'name'} <= set(updated_party_data):
+            model_result = PartiesModelDb(party_id=oid).edit_party(updated_party_data['name'])
+            if 'Invalid Id' in model_result or 'Invalid Data' in model_result:
+                return make_response(jsonify({"status": 400, "error": "Invalid Data ,Check id or data being updated"}),
+                                     400)
+            elif 'Party Exists' in model_result:
+                return make_response(jsonify({"status": 409, "error": "Party with similar name exists"}), 409)
+            return make_response(
+                jsonify({"status": 200, "message": "{} Updated Successfully".format(model_result[0][1])}),
+                200)
+        return make_response(jsonify({"status": 400, "error": "Missing Key value"}), 400)
+    return make_response(jsonify({"status": 401, "error": "Unauthorized Access,Requires Admin Rights"}), 401)
 
 
 @parties_api_v2.route("/parties/<party_id>", methods=['GET'])
 @jwt_required
-def api_specific_office_get(party_id):
+def api_specific_party_get(party_id):
     oid = id_conversion(party_id)
     party = PartiesModelDb(party_id=oid).get_specific_party()
     if isinstance(party, list) and len(party) >= 1:
@@ -67,12 +78,16 @@ def api_specific_office_get(party_id):
 
 @parties_api_v2.route("/parties/<party_id>", methods=['DELETE'])
 @jwt_required
-def api_specific_office_delete(party_id):
-    oid = id_conversion(party_id)
-    party = PartiesModelDb(party_id=oid).delete_party()
-    if isinstance(party, list):
-        return make_response(jsonify({"status": 200, "message": "{} Deleted".format(party[0][0])}), 200)
-    return make_response(jsonify({"status": 404, "error": "Data Not Found"}), 404)
+def api_specific_party_delete(party_id):
+    current_user = get_jwt_identity()
+    response = UserModelDb().get_user_by_id(current_user)
+    if 'Requires Admin Privilege' not in response:
+        oid = id_conversion(party_id)
+        party = PartiesModelDb(party_id=oid).delete_party()
+        if isinstance(party, list):
+            return make_response(jsonify({"status": 200, "message": "{} Deleted".format(party[0][0])}), 200)
+        return make_response(jsonify({"status": 404, "error": "Data Not Found"}), 404)
+    return make_response(jsonify({"status": 401, "error": "Unauthorized Access,Requires Admin Rights"}), 401)
 
 
 def id_conversion(item_id):
